\name{Original_ex}
\alias{Original_ex}
\docType{data}
\title{
An hemispherical picture
}
\description{
This dataset is an example of a typical hemispherical picture.
}
\usage{data(Original_ex)}
\format{
  The format is:
Formal class 'Image' [package "EBImage"] with 2 slots
  ..@ .Data    : num [1:600, 1:450, 1:3] 0 0 0 0 0 0 0 0 0 0 ...
  ..@ colormode: int 2
}
\examples{
data(Original_ex)
display(Original_ex)
}
\keyword{datasets}
